requirements.txt
streamlit>=1.33
python-dotenv>=1.0.1

# Document parsing
pypdf>=4.0.0
python-docx>=1.1.0

# RAG
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0

# Embeddings
sentence-transformers>=2.2.0

# LLM providers (choose one via config)
langchain-openai>=0.1.0
langchain-ollama>=0.1.0

# Database operations
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0
pyodbc>=4.0.0

# System monitoring
psutil>=5.9.0
schedule>=1.2.0

# Machine learning for predictive analysis
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

